# openapi.yaml
openapi: 3.0.3
info:
  title: Clair API
  version: 1.0.1
  description: |
    API documentation for Clair backend based on the current project routes.
    Includes Gemini key save/replace and delete.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    channelKeyAuth:
      type: apiKey
      in: header
      name: x-channel-key
  schemas:
...
    CreateChannelBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Main Site
        allowed_domain:
          type: string
          example: mysite.com
...
    InternalAppealBody:
      type: object
      required: [cid, text]
      properties:
        cid:
          type: integer
          example: 12
        text:
          type: string
          example: Не работает форма оплаты
...
    AiTextBody:
      type: object
      required: [text]
      properties:
        text:
          type: string
          example: Hello, I need help with checkout
    RabbitTestBody:
      type: object
      properties:
        cid:
          type: integer
          example: 12
        foo:
          type: string
          example: bar
paths:
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
        "401":
          description: Unauthorized
...
  /api/me:
    get:
      tags: [Auth]
      summary: Test protected route
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current req.user
        "401":
          description: Unauthorized
...
  /api/channels:
    get:
      tags: [Channels]
      summary: Get all channels of current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Channel list
        "401":
          description: Unauthorized
    post:
      tags: [Channels]
      summary: Create channel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelBody"
      responses:
        "201":
          description: Channel created
        "401":
          description: Unauthorized
        "400":
          description: Bad request
...
  /api/appeals:
    post:
      tags: [Appeals]
      summary: Create internal appeal and push to channel queue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalAppealBody"
      responses:
        "202":
          description: Appeal queued
        "400":
          description: Bad request
        "401":
          description: Unauthorized
...
  /api/context:
    post:
      tags: [AI]
      summary: Generic AI text response
      description: |
        Controller expects req.user.id. This route should be protected by JWT in your project setup.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiTextBody"
      responses:
        "200":
          description: AI response
        "401":
          description: Unauthorized
...
  /api/rabbit/test:
    post:
      tags: [Rabbit Test]
      summary: Push test message to RabbitMQ
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RabbitTestBody"
      responses:
        "200":
          description: Test message queued